requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-environ==0.10.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.19.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
